# PERF: MultiIndex._shallow_copy