--- a/pandas/tests/indexes/common.py
+++ b/pandas/tests/indexes/common.py
@@ -923,11 +923,12 @@ class Base:
     def test_shallow_copy_copies_cache(self):
         # GH32669
         idx = self.create_index()
-        idx.get_loc(idx[0])  # initiate the _engine etc.
+        idx.get_loc(idx[0])  # populates the _cache.
         shallow_copy = idx._shallow_copy()
+
         # check that the shallow_copied cache is a copy of the original
         assert idx._cache == shallow_copy._cache
         assert idx._cache is not shallow_copy._cache
-        # the cache values should reference the same objects
+        # cache values should reference the same object
         for key, val in idx._cache.items():
-            assert shallow_copy._cache[key] is val
+            assert shallow_copy._cache[key] is val, key