@topper-123 2020-03-13
It was stupid to do ops in the "levels" in the cache. Simpler to delete "levels" and just let it be recreated if/when the .levels attribute is accessed in the new shallow_copied index.