@ja3067 2020-03-28
This PR has introduced a small bug. I'm not sure how deep it goes, but I've encountered it in `PeriodIndex` objects inside a `MultiIndex`. Here is a minimal reproduction:

```python
import pandas as pd
idx = pd.MultiIndex.from_arrays([pd.PeriodIndex([pd.Period("2019Q1"), pd.Period("2019Q2")], name='b')])
idx2 = pd.MultiIndex.from_arrays([idx._get_level_values(level) for level in range(idx.nlevels)])
print(all(x.is_monotonic for x in idx2.levels)) # raises an error
```

There's a simple fix for this, which involves changing `pandas/core/indexes/period.py:260` to `result._cache = {}`. I don't understand the internals enough to know what is causing this issue or whether it extends farther than this case. The problem seems to be that the `_cache` caches the `_engine` attribute, and since we no longer clear the cache when copying the object using `_simple_new`, `_engine` is left in an invalid state. The error is caused by `idx._simple_new()._engine.vgetter()` returning None, which causes issues in `is_monotonic`, among other things.