@topper-123 2020-03-29
Thanks for the report, @jacobaustin123. This is a bit tricky as you mention, but is caused by the use of ``weakref(self)`` is ``PeriodIndex._engine``, where the weakreffed self in this case is released too early.

This can be solved by deleting ``PeriodIndex._engine`` and falling back to the base ``Index._engine``.

@jbrockmendel , I think you made the new ``Index._engine`` method including ``Index._get_engine_target``. Do you agree that ``PeriodIndex._engine`` should just be deleted?