@jbrockmendel 2020-03-29
> Do you agree that PeriodIndex._engine should just be deleted?

Eventually, yes.  The reason why PeriodIndex._engine exists is because PeriodEngine needs to get the `.freq` attribute of the PeriodIndex.  However, the PeriodEngine methods that actually use `.freq` are never actually reached, because the relevant PeriodIndex methods route around them, and it isn't obvious how intentional that is.

So it should be removed, but there is some untangling to do first.