@topper-123 2020-03-29
Ok. a fix to the current problem seems to be to replace ``weakref(self)`` with ``weakref(self._values)`` (haven't run the test suite. We could do that until the fix, I think.